Fix 32-bit addresses
Because high 16 bits were lost during lexing. Also changed memory size

+++ include/ast.h
@@ -59,7 +59,7 @@ typedef struct {
 } Argument;
 
 typedef struct {
-    unsigned int address;
+    unsigned long int address;
     bool isMacro;
     union {
         enum instrType iType;
